% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmn.R
\name{char2intvect}
\alias{char2intvect}
\title{Converts model.names from character to numeric representations}
\usage{
char2intvect(charnames)
}
\arguments{
\item{charnames}{A character vector, should represent in character form numeric vectors concisting of the integers 2:n. Formats such as c(2,4) and 2:5 are allowed.}
}
\value{
\code{char2intvect} returns a list of character vectors representing the same information. Crucial: callng \code{char2intvect} on the output of \code{intvect2char} should always get you back to the original starting input.
}
\description{
Inverse conversion of intvect2char.
}
