% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errcheck_mats.R
\name{errcheck_mats}
\alias{errcheck_mats}
\title{Error checking for the input \code{mats}
 
Error checking for the input \code{mats}}
\usage{
errcheck_mats(comingfrom, mats)
}
\arguments{
\item{comingfrom}{Name of the function calling this one}

\item{mats}{The argument \code{mats} to \code{mmsscore}, \code{mmsrank}, \code{mmsmodwts}, \code{matregtest}}
}
\value{
\code{errcheck_mats} Returns nothing. Throws and error if there is a problem with inputs.
}
\description{
Error checking for the input \code{mats}
 
Error checking for the input \code{mats}
}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
