% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errcheck_n.R
\name{errcheck_n}
\alias{errcheck_n}
\title{Error checking for the input \code{n} and \code{maxruns}
 
Error checking for the input \code{n} and \code{maxruns}}
\usage{
errcheck_n(comingfrom, dimmats, n, maxruns)
}
\arguments{
\item{comingfrom}{Name of the function calling this one}

\item{dimmats}{The dimension of the \code{mats} argument that is passed in to \code{matregtest}, \code{mmsscore}, \code{mmsrank}, \code{mmsmodwts}}

\item{n}{The argument \code{n} to \code{mmsscore}, \code{mmsrank}, \code{mmsmodwts}}

\item{maxruns}{The argument \code{maxruns} to \code{mmsscore}, \code{mmsrank}, \code{mmsmodwts}}
}
\value{
\code{errcheck_n} Returns nothing. Throws and error if there is a problem with inputs.
}
\description{
Error checking for the input \code{n} and \code{maxruns}
 
Error checking for the input \code{n} and \code{maxruns}
}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
