% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errcheck_pred.R
\name{errcheck_pred}
\alias{errcheck_pred}
\title{Error checking for the inputs \code{pred} and \code{model.names}}
\usage{
errcheck_pred(comingfrom, predlist, lenmats)
}
\arguments{
\item{comingfrom}{Name of the function calling this one}

\item{predlist}{Either \code{list(pred)} or \code{model.names}}

\item{lenmats}{\code{length(mats)}}
}
\value{
\code{errcheck_pred} Returns nothing. Throws and error if there is a problem with inputs.
}
\description{
Error checking for the inputs \code{pred} and \code{model.names}
}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
