% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmn.R
\name{intvect2char}
\alias{intvect2char}
\title{Converts model.names from numeric to character representations}
\usage{
intvect2char(numnames)
}
\arguments{
\item{numnames}{A list of numeric vectors each consisting of the numbers \code{2:n}}
}
\value{
\code{intvect2char} returns a character vector. Crucial: callng \code{char2intvect} on the output of \code{intvect2char} should always get you back to the original starting input.
}
\description{
Converts a list of vectors consisting of the numbers 2:n into a character
vector representing the same thing.
}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
