% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kyldisp}
\alias{kyldisp}
\title{Qualitatively estimated dispersal difficulty matrix for Kentucky Lake}
\format{A numeric matrix with rownames and colnames containing sampling location names}
\usage{
kyldisp
}
\description{
This matrix was created and analyzed by Anderson et al. (Oikos, 127, 403-414, 2017) and was provided
in the supporting information of that paper. It is reproduced here with no alteration.
The matrix represents qualitative estimates of the difficulty of plankton dispersal between all pairs
of sampling locations, with larger numbers meaning more difficult dispersal.
Estimates were based on knowledge of large-scale water movement patterns in the lake.
}
\keyword{datasets}
