% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makenames.R
\name{makenames}
\alias{makenames}
\title{Makes model names}
\usage{
makenames(lenmats)
}
\arguments{
\item{lenmats}{The number of matrices in the \code{mats} argument provided to various other functions}
}
\value{
\code{makenames} A list of numeric vectors containing the number 2:lenmats
}
\description{
Makes model names. These are all combinations of the integers \code{2:lenmats}
}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
