% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmsclean.R
\name{mmsclean}
\alias{mmsclean}
\title{Clean data for model selection for matrix models}
\usage{
mmsclean(mats, resp, pred, n)
}
\arguments{
\item{mats}{A named list of matrices, all assumed to be the same dimensions. Only the lower triangles are used (not including the diagonal). NA/NaNs are allowed. (See \code{\link{table2matrix}}.)}

\item{resp}{The index in mats of the response variable (input is a numeric value, e.g. resp = 1)}

\item{pred}{The indices in mats of predictor variables, should not include resp. Input is numeric value(s), e.g. pred=1, pred =1:2, pred =c(1,2,4).}

\item{n}{The number of sampling locations to leave out. Must be at least 2.}
}
\value{
\code{mmsclean} return an object of class list consisting of 
\item{mat}{List of matrices containing response and predictors}
\item{resp}{Index of response variable}
\item{pred}{Indices of predictor variables}
}
\description{
This function is used to clean data prior to input into leave-n-out cross valdation (LNOCV) model selection procedures
}
\examples{
x<-list(resp=matrix(rnorm(100),nrow=10,ncol=10),pred1=matrix(rnorm(100),nrow=10,ncol=10),pred2=matrix(rnorm(100),nrow=10,ncol=10))
y<-mmsclean(mats=x,pred=1,resp=2:3,n=3)
 
}
\author{
Tom Anderson, \email{anderstl@gmail.edu}; Daniel Reuman, \email{reuman@ku.edu}; Jon Walter, \email{jaw3es@virginia.edu}
}
\keyword{internal}
