% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmsscore_int.R
\name{mmsscore_int}
\alias{mmsscore_int}
\title{Calculate leave-n-out cross validation (LNOCV) score}
\usage{
mmsscore_int(mats, pred, n, maxruns)
}
\arguments{
\item{mats}{A list of matrices, all assumed to be the same dimensions. Only the lower triangles are used. NA/NaNs are allowed. The first entry taken to be the response.}

\item{pred}{The indices in mats of predictor variables in the more complex of the two models to be compared, should not include 1. Input is numeric value(s), e.g. pred=2, pred =2:3, pred =c(2,3,5).}

\item{n}{The number of sampling locations to leave out, must be at least 2.}

\item{maxruns}{The maximum number of leave-n-outs (LNOs) to do. To be used if choose(dim(mats[[1]]),n) is very large. Inf to use (or try to use) all LNOs. If maxruns is a number, then LNOs are selected randomly and hence may include repeats.}
}
\value{
\code{mmsscore} return an object of class list consisting of 
\item{lno.score}{The out-of-sample forecast accuracy (mean squared error)}
\item{num.pos}{The possible number of LNOs for the given n and number of locations}
\item{num.att}{The total number of LNOs attempted}
\item{num.rnk}{The number of LNOs that did not result in a rank deficiency regression problem, and so could be used for testing out-of-sample predictions}
\item{num.usd}{The number of LNOs that could be used in the end (possibly less than num.rnk because of NAs in the input matrices)}
}
\description{
This function is used to calculate an LNOCV score for a given model, based on mean squared error, out of sample. Same as mmsscore but with no error checking.
}
\author{
Tom Anderson, \email{anderstl@gmail.edu}; Daniel Reuman, \email{reuman@ku.edu}; Jon Walter, \email{jaw3es@virginia.edu}
}
\keyword{internal}
