% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmsvarwts.R
\name{mmsvarwts}
\alias{mmsvarwts}
\title{Variable importance}
\usage{
mmsvarwts(pred, weights, varnames = NULL)
}
\arguments{
\item{pred}{a numeric vector giving indices of predictor variables}

\item{weights}{a data frame of model weights, output from \code{mmsmodwts}}

\item{varnames}{a character vector of predictor names. If NULL (the default), the indices in \code{pred} are used as names}
}
\value{
an object of class dataframe consisting of summed variable weights
}
\description{
Function to calculate variable importance by summing LNOCV weights.
}
\examples{
x<-list(resp=matrix(rnorm(100),nrow=10,ncol=10),pred1=matrix(rnorm(100),nrow=10,ncol=10),pred2=matrix(rnorm(100),nrow=10,ncol=10),pred3=matrix(rnorm(100),nrow=10,ncol=10))
y<-mmsmodwts(x,model.names=NA,n=3,maxruns=10,nrand=10) 
mmsvarwts(pred=2:4,weights=y,varnames=names(x))

}
\author{
Jon Walter, \email{jaw3es@virginia.edu}
}
