% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmn.R
\name{transmn}
\alias{transmn}
\title{Translates models.names from character to numeric representations or the reverse}
\usage{
transmn(mn, inform)
}
\arguments{
\item{mn}{The model names in one format or the other}

\item{inform}{Either "char" or "numvect"}
}
\value{
\code{transmn} returns the other format
}
\description{
Translates models.names from character to numeric representations or the reverse
}
