% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kyldat}
\alias{kyldat}
\title{Data from Kentucky Lake}
\format{An object of class \code{data.frame} with 416 rows and 27 columns.}
\usage{
kyldat
}
\description{
Plankton and environmental measurements from a long term monitoring project on Kentucky
Lake, Kentucky, USA.
These are data analyzed by Anderson et al. (Oikos, 127, 403-414, 2017) and stored in Dryad
by the same authors (Dryad Digital Repository, https://doi.org/10.5061/dryad.21jt3, 2017), provided
here with no alteration.
Data are annual averages from 16 lake locations from 1990-2015. 
The original, pre-annualization data can be requested at http://www.murraystate.edu/wsi/wsi_database.aspx.
}
\keyword{datasets}
