% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeform.R
\name{makeform}
\alias{makeform}
\title{Makes a formula for regression models}
\usage{
makeform(mats)
}
\arguments{
\item{mats}{The argument \code{mats} to various other functions in the package}
}
\value{
\code{makeform} A regression model formula using the names of \code{mats}
}
\description{
Makes a formula for regression models
}
\author{
Daniel Reuman, \email{reuman@ku.edu}
}
