% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matregtest.R
\name{matregtest}
\alias{matregtest}
\title{Statistically compares two nested matrix models.}
\usage{
matregtest(mats, resp, pred, drop, numperm)
}
\arguments{
\item{mats}{A named list of matrices, all assumed to be the same dimensions. Only the lower triangles are used (not including the diagonal). NA/NaNs are allowed. (See \code{\link{table2matrix}}.)}

\item{resp}{The index in mats of the response variable (input is a numeric value, e.g. resp = 1)}

\item{pred}{The indices in mats of predictor variables in the more complex of the two models to be compared, should not include resp. Input is numeric value(s), e.g. pred=1, pred =1:2, pred =c(1,2,4).}

\item{drop}{The indices in mats of predictor variables that are dropped to get from the complex to the simple model. Should be a subset of pred.}

\item{numperm}{The number of permutations used to do the test}
}
\value{
\code{matregtest} return an object of class list consisting of 
\item{ssr_dat}{Sum of squared residuals for the linear regression using all predictors in \code{pred}}
\item{ssr_perm}{Vector of sums of squares of residuals for linear regressions using randomized matrices for the indices in \code{drop}}
\item{p}{The approximate p-value for the test with null hypothesis the simpler model and alternative the more complex one}
}
\description{
A function for doing matrix regression tests of a model against a nested model.
}
\examples{

}
\author{
Lei Zhao, \email{leizhao@ku.edu} ; Daniel Reuman, \email{reuman@ku.edu}
}
