% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmsvarwts.R
\name{mmsvarwts}
\alias{mmsvarwts}
\title{Variable importance}
\usage{
mmsvarwts(pred, weights, varnames = NULL)
}
\arguments{
\item{pred}{a numeric vector giving indices of predictor variables}

\item{weights}{a data frame of model weights, output from \code{mmsmodwts}}

\item{varnames}{a character vector of predictor names. If NULL (the default), the indices in \code{pred} are used as names}
}
\value{
an object of class dataframe consisting of summed variable weights
}
\description{
Function to calculate variable importance by summing LNOCV weights.
}
\examples{

}
\author{
Jon Walter, \email{jaw3es@virginia.edu}
}
